# docs/swagger.yaml
openapi: 3.0.3
info:
  title: shb API
  version: 1.0.0
  contact:
    name: shb Support

servers:
  - url: http://localhost:8000/api/v1
    description: Сервер разработки
  - url: https://localhost:8000
    description: Root server

tags:
  - name: Health
    description: Проверка состояния сервиса
  - name: Institutions
    description: Учреждения
  - name: Events
    description: События
  - name: Campaigns
    description: Медицинские кампании и прозрачность пожертвований
  - name: Donations
    description: Пожертвования пользователя

paths:
  /institutions:
    get:
      tags:
        - Institutions
      summary: Список учреждений (пагинация)
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Поиск по названию или городу
        - name: type
          in: query
          schema:
            type: string
          description: Тип учреждения или all
        - name: sort
          in: query
          schema:
            type: string
            enum: [needs_desc, distance]
          description: Сортировка
        - name: lat
          in: query
          schema:
            type: number
            format: double
        - name: lng
          in: query
          schema:
            type: number
            format: double
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Размер страницы (по умолчанию 20, макс. 100)
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Страница списка
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/InstitutionPage'
        '400':
          description: Некорректные limit/offset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

  /events:
    get:
      tags:
        - Events
      summary: Список событий (пагинация)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Страница списка
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/EventPage'
        '400':
          description: Некорректные limit/offset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

  /events/{id}:
    get:
      tags:
        - Events
      summary: Событие по ID
      description: |
        Опциональный Bearer — для поля is_joined в теле ответа.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      security:
        - {}
        - BearerAuth: []
      responses:
        '200':
          description: Карточка события (как элемент списка)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/EventResponseItem'
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '404':
          description: Событие не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  /campaigns:
    get:
      tags:
        - Campaigns
      summary: Список активных кампаний
      security: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Страница активных кампаний
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/CampaignPage'
        '400':
          description: Некорректные limit/offset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'

  /campaigns/{id}:
    get:
      tags:
        - Campaigns
      summary: Детали кампании
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Детали кампании без банковских реквизитов клиники
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/PublicCampaign'
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '404':
          description: Кампания не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  /campaigns/{id}/ledger:
    get:
      tags:
        - Campaigns
      summary: Финансовый ledger кампании
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Транзакции кампании
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LedgerEntryPage'
        '400':
          description: Некорректный id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '404':
          description: Кампания не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundError'

  /campaigns/{id}/donate:
    post:
      tags:
        - Donations
      summary: Пожертвовать в кампанию
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DonationRequest'
      responses:
        '200':
          description: Пожертвование записано
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LedgerEntry'
        '400':
          description: Некорректная сумма или id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestError'
        '401':
          description: Требуется авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

  /donations/my:
    get:
      tags:
        - Donations
      summary: Мои пожертвования
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Пожертвования текущего пользователя
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/LedgerEntryPage'
        '401':
          description: Требуется авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnauthorizedError'

  /ping:
    get:
      tags:
        - Health
      summary: Проверка состояния сервиса
      description: Простой эндпоинт для проверки работоспособности API
      responses:
        '200':
          description: Сервис работает
          content:
            text/plain:
              schema:
                type: string
                example: "pong"
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT токен доступа. Добавьте токен в заголовок Authorization в формате:
        `Authorization: Bearer <token>`

  schemas:
    SuccessResponse:
      type: object
      properties:
        message:
          type: string
          example: "Success"
        payload:
          type: object
          description: Данные ответа
      required:
        - message

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "service temporarily unavailable"

    BadRequestError:
      type: object
      properties:
        message:
          type: string
          example: "invalid request body"

    PublicCampaign:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        target_amount:
          type: number
          format: double
        collected_amount:
          type: number
          format: double
        currency:
          type: string
        status:
          type: string
        deadline:
          type: string
          format: date-time
          nullable: true
        completed_at:
          type: string
          format: date-time
          nullable: true
        cancelled_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
        clinic_name:
          type: string
          nullable: true

    CampaignPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PublicCampaign'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
        campaign_id:
          type: integer
          nullable: true
        type:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
        description:
          type: string
          nullable: true
        donor_name:
          type: string
          nullable: true
        is_anonymous:
          type: boolean
        donor_message:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    LedgerEntryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LedgerEntry'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    DonationRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true

    NotFoundError:
      type: object
      properties:
        message:
          type: string
          example: "not found"

    EventResponseItem:
      type: object
      description: Элемент списка / карточка события
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        event_date:
          type: string
          format: date-time
        institution_id:
          type: integer
        institution_name:
          type: string
        creator_id:
          type: integer
        creator_name:
          type: string
        participants_count:
          type: integer
        is_joined:
          type: boolean
        created_at:
          type: string
          format: date-time

    UnauthorizedError:
      type: object
      properties:
        message:
          type: string
          example: "invalid OTP"

    ForbiddenError:
      type: object
      properties:
        message:
          type: string
          example: "access denied"

    UnprocessableError:
      type: object
      properties:
        message:
          type: string
          example: "validation failed"

    TooManyRequestsError:
      type: object
      properties:
        message:
          type: string
          example: "phone number is temporarily blocked due to too many requests"

    InstitutionPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset

    EventPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
        total:
          type: integer
          format: int64
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset

security:
  - BearerAuth: []